+++ src/__test__/components/DeployContract/DeployContractForm.test.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import DeployContractForm from './DeployContractForm';
-import QuickDeployment from './QuickDeployment';
-import GuidedDeployment from './GuidedDeployment';
+import DeployContractForm from '../../../components/DeployContract/DeployContractForm';
+import QuickDeployment from '../../../components/DeployContract/QuickDeployment';
+import GuidedDeployment from '../../../components/DeployContract/GuidedDeployment';
 
 describe('DeployContractForm', () => {
   let deployContractForm;
@@ -12,8 +12,8 @@ describe('DeployContractForm', () => {
   beforeEach(() => {
     const props = {
       location: {}
-    }
-    deployContractForm = shallow(<DeployContractForm {...props}/>)
+    };
+    deployContractForm = shallow(<DeployContractForm {...props}/>);
   });
   
   it('should render QuickDeployment by default', () => {
@@ -21,25 +21,25 @@ describe('DeployContractForm', () => {
       location: {
         search: ""
       }
-    })
+    });
     expect(deployContractForm.find(QuickDeployment)).to.have.length(1);
-  })
+  });
 
   it('should render QuickDeployment if mode is quick', () => {
     deployContractForm.setProps({
       location: {
         search: "?mode=quick"
       }
-    })
+    });
     expect(deployContractForm.find(QuickDeployment)).to.have.length(1);
-  })
+  });
 
   it('should render GuidedDeployment if mode is guided', () => {
     deployContractForm.setProps({
       location: {
         search: "?mode=guided"
       }
-    })
+    });
     expect(deployContractForm.find(GuidedDeployment)).to.have.length(1);
-  })
-})
+  });
+});
