+++ src/__test__/components/DeployContract/GuidedDeployment.test.js
@@ -4,16 +4,16 @@ import { Steps } from 'antd';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import GuidedDeployment from './GuidedDeployment';
+import GuidedDeployment from '../../../components/DeployContract/GuidedDeployment';
 
 describe('GuidedDeployment', () => {
   let guidedDeployment;
 
   beforeEach(() => {
-    guidedDeployment = shallow(<GuidedDeployment />)
+    guidedDeployment = shallow(<GuidedDeployment />);
   });
   
   it('should render 4 steps', () => {
     expect(guidedDeployment.find(Steps.Step)).to.have.length(5);
-  })
-})
+  });
+});
