+++ src/__test__/components/DeployContract/QuickDeployment.test.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 
-import QuickDeployment from './QuickDeployment';
+import QuickDeployment from '../../../components/DeployContract/QuickDeployment';
 
 describe('QuickDeployment', () => {
   let quickDeployment;
   
   it('should render', () => {
-    quickDeployment = mount(<QuickDeployment />)
-  })
-})
+    quickDeployment = mount(<QuickDeployment />);
+  });
+});
